start to unify thread state
