+++ lib/asmrepl/linux.rb
@@ -1,4 +1,5 @@
 require "fisk/helpers"
+require "asmrepl/thread_state"
 
 module ASMREPL
   module Linux
@@ -50,8 +51,7 @@ module ASMREPL
       raise unless ptrace(PTRACE_TRACEME, 0, 0, 0).zero?
     end
 
-    class ThreadState
-      fields = (<<-eostruct).scan(/int ([^;]*);/).flatten
+    fields = (<<-eostruct).scan(/int ([^;]*);/).flatten
 struct user_regs_struct
 {
   __extension__ unsigned long long int r15;
@@ -82,92 +82,15 @@ struct user_regs_struct
   __extension__ unsigned long long int fs;
   __extension__ unsigned long long int gs;
 };
-      eostruct
-      fields.each_with_index do |field, i|
-        define_method(field) do
-          to_ptr[Fiddle::SIZEOF_INT64_T * i, Fiddle::SIZEOF_INT64_T].unpack1("l!")
-        end
+    eostruct
 
-        define_method("#{field}=") do |v|
-          to_ptr[Fiddle::SIZEOF_INT64_T * i, Fiddle::SIZEOF_INT64_T] = [v].pack("l!")
-        end
-      end
+    class ThreadState < ASMREPL::ThreadState.build(fields)
+      private
 
-      define_singleton_method(:sizeof) do
-        fields.length * Fiddle::SIZEOF_INT64_T
-      end
+      def read_flags; eflags; end
 
-      def [] name
-        idx = fields.index(name)
-        return unless idx
-        to_ptr[Fiddle::SIZEOF_INT64_T * idx, Fiddle::SIZEOF_INT64_T].unpack1("l!")
-      end
-
-      def []= name, val
-        idx = fields.index(name)
-        return unless idx
-        to_ptr[Fiddle::SIZEOF_INT64_T * idx, Fiddle::SIZEOF_INT64_T] = [val].pack("l!")
-      end
-
-      def self.malloc
-        new Fiddle::Pointer.malloc sizeof
-      end
-
-      attr_reader :to_ptr
-
-      def initialize buffer
-        @to_ptr = buffer
-      end
-
-      define_method(:fields) do
-        fields
-      end
-
-      def to_s
-        buf = ""
-        fields.first(8).zip(fields.drop(8).first(8)).each do |l, r|
-          buf << "#{l.ljust(3)}  #{sprintf("%#018x", send(l))}"
-          buf << "  "
-          buf << "#{r.ljust(3)}  #{sprintf("%#018x", send(r))}\n"
-        end
-
-        buf << "\n"
-
-        fields.drop(16).each do |reg|
-          buf << "#{reg.ljust(8)}  #{sprintf("%#018x", send(reg))}\n"
-        end
-        buf
-      end
-
-      FLAGS = [
-        ['CF', 'Carry Flag'],
-        [nil, 'Reserved'],
-        ['PF', 'Parity Flag'],
-        [nil, 'Reserved'],
-        ['AF', 'Adjust Flag'],
-        [nil, 'Reserved'],
-        ['ZF', 'Zero Flag'],
-        ['SF', 'Sign Flag'],
-        ['TF', 'Trap Flag'],
-        ['IF', 'Interrupt Enable Flag'],
-        ['DF', 'Direction Flag'],
-        ['OF', 'Overflow Flag'],
-        ['IOPL_H', 'I/O privilege level High bit'],
-        ['IOPL_L', 'I/O privilege level Low bit'],
-        ['NT', 'Nested Task Flag'],
-        [nil, 'Reserved'],
-      ]
-
-      def flags
-        flags = eflags
-        f = []
-        FLAGS.each do |abbrv, _|
-          if abbrv && flags & 1 == 1
-            f << abbrv
-          end
-          flags >>= 1
-        end
-        f
+      def other_registers
+        super - ["orig_rax"]
       end
     end
